llm_service/requirements.txt
# --- Core Framework ---
fastapi==0.118.1
uvicorn[standard]==0.37.0

# --- Config / Validation ---
pydantic==2.12.0
pydantic-settings==2.11.0
python-dotenv==1.1.1

# --- AI / LangGraph Stack ---
langgraph==0.6.9
langchain==0.3.27
langchain-openai==0.3.35
openai==2.2.0

# --- Auth / Security ---
PyJWT==2.10.1

# --- Network / External API ---
requests==2.32.5
ytmusicapi==1.11.1
